Chat-Scrapper/requirements.txt
requests
beautifulsoup4
faiss-cpu
numpy
python-dotenv